test(consumer): harden crash harness cleanup
Refs #1615
