+++ tests/Dekaf.Tests.Integration/TransactionalCrashClientProcess.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using System.Text;
 using Dekaf.Consumer;
 using Dekaf.Producer;
 using Dekaf.Protocol.Messages;
@@ -300,35 +299,6 @@ internal sealed class TransactionalCrashClientProcess : IAsyncDisposable
         }
     }
 
-    private sealed class BoundedProcessDiagnostics(int maximumCharacters)
-    {
-        private readonly StringBuilder _buffer = new();
-
-        public void Append(string stream, string? message)
-        {
-            if (message is null)
-            {
-                return;
-            }
-
-            lock (_buffer)
-            {
-                _buffer.Append('[').Append(stream).Append("] ").AppendLine(message);
-                if (_buffer.Length > maximumCharacters)
-                {
-                    _buffer.Remove(0, _buffer.Length - maximumCharacters);
-                }
-            }
-        }
-
-        public override string ToString()
-        {
-            lock (_buffer)
-            {
-                return _buffer.ToString();
-            }
-        }
-    }
 }
 
 /// <summary>
